requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
-e git+https://github.com/schmidtbri/regression-model#egg=insurance_charges_model
    # via -r requirements.in
anyio==3.6.1
    # via starlette
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   distributed
    #   featuretools
    #   uvicorn
cloudpickle==2.1.0
    # via
    #   dask
    #   distributed
    #   featuretools
dask[dataframe]==2022.7.1
    # via
    #   distributed
    #   featuretools
deap==1.3.1
    # via tpot
debugpy==1.6.2
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
deprecated==1.2.13
    # via redis
distributed==2022.7.1
    # via featuretools
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.9.1
    # via stack-data
faker==13.15.1
    # via -r requirements.in
fastapi==0.79.0
    # via rest-model-service
fastjsonschema==2.16.1
    # via nbformat
featuretools==0.24.0
    # via insurance-charges-model
fsspec==2022.7.1
    # via dask
h11==0.13.0
    # via uvicorn
heapdict==1.0.1
    # via zict
idna==3.3
    # via
    #   anyio
    #   requests
ipykernel==6.15.1
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.4.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
    #   qtconsole
ipywidgets==7.7.1
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   distributed
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via
    #   scikit-learn
    #   tpot
jsonschema==4.9.1
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.3.4
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.1
    # via ipywidgets
locket==1.0.0
    # via
    #   distributed
    #   partd
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
ml-base==0.2.1
    # via
    #   -r requirements.in
    #   insurance-charges-model
    #   rest-model-service
msgpack==1.0.4
    # via
    #   -r requirements.in
    #   distributed
nbclient==0.6.6
    # via nbconvert
nbconvert==6.5.0
    # via
    #   jupyter
    #   notebook
nbformat==5.4.0
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.23.1
    # via
    #   dask
    #   deap
    #   featuretools
    #   pandas
    #   scikit-learn
    #   scipy
    #   tpot
    #   xgboost
packaging==21.3
    # via
    #   dask
    #   distributed
    #   ipykernel
    #   nbconvert
    #   qtpy
    #   redis
pandas==1.4.3
    # via
    #   dask
    #   featuretools
    #   insurance-charges-model
    #   tpot
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
partd==1.2.0
    # via dask
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.30
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.1
    # via
    #   distributed
    #   featuretools
    #   ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via
    #   fastapi
    #   ml-base
pygments==2.12.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   faker
    #   jupyter-client
    #   pandas
python-snappy==0.6.1
    # via -r requirements.in
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via
    #   dask
    #   distributed
    #   featuretools
    #   rest-model-service
pyzmq==23.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.1
    # via jupyter
qtpy==2.1.0
    # via qtconsole
redis==4.3.4
    # via -r requirements.in
requests==2.28.1
    # via update-checker
rest-model-service==0.3.1
    # via
    #   -r requirements.in
    #   insurance-charges-model
scikit-learn==0.24.2
    # via
    #   insurance-charges-model
    #   tpot
scipy==1.9.0
    # via
    #   featuretools
    #   scikit-learn
    #   tpot
    #   xgboost
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
sniffio==1.2.0
    # via anyio
sortedcontainers==2.4.0
    # via distributed
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.3.0
    # via ipython
starlette==0.19.1
    # via fastapi
stopit==1.1.2
    # via tpot
tblib==1.7.0
    # via distributed
terminado==0.15.0
    # via notebook
threadpoolctl==3.1.0
    # via scikit-learn
tinycss2==1.1.1
    # via nbconvert
toolz==0.12.0
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via
    #   distributed
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tpot==0.11.7
    # via insurance-charges-model
tqdm==4.64.0
    # via
    #   featuretools
    #   tpot
traitlets==5.3.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
update-checker==0.18.0
    # via tpot
urllib3==1.26.11
    # via
    #   distributed
    #   requests
uvicorn==0.18.2
    # via rest-model-service
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==3.6.1
    # via ipywidgets
wrapt==1.14.1
    # via deprecated
xgboost==1.6.1
    # via tpot
zict==2.2.0
    # via distributed

# The following packages are considered to be unsafe in a requirements file:
# setuptools
